Move `TApp` hooking code to new file
