+++ srcDLL/DllMain.cpp
@@ -1,4 +1,5 @@
 #include "Globals.h"
+#include "AppEvents.h"
 #include "ModuleLoader.h"
 #include "VolList.h"
 #include "StringConversion.h"
@@ -31,6 +32,7 @@ LoggerMessageBox loggerMessageBox; // Logging to pop-up message box
 LoggerDistributor loggerDistributor({&loggerFile, &loggerMessageBox});
 LoggerDebug loggerDebug;
 std::unique_ptr<VolList> volList;
+AppEvents appEvents;
 
 
 BOOL WINAPI DllMain(HINSTANCE hInstance, DWORD dwReason, LPVOID reserved)
@@ -56,6 +58,14 @@ BOOL WINAPI DllMain(HINSTANCE hInstance, DWORD dwReason, LPVOID reserved)
 				LogError("Failed to install initial TApp event hooks. Module loading and patching disabled.");
 				return FALSE;
 			}
+
+			// Setup event handlers
+			appEvents.onInit.Add(&OnInit);
+			appEvents.onLoadShell.Add(&OnLoadShell);
+			appEvents.onShutDown.Add(&OnShutdown);
+
+			// Set active events
+			appEvents.Activate();
 		}
 
 		// Disable any more thread attach calls
@@ -128,71 +138,4 @@ void OnLoadShell()
 void OnShutdown()
 {
 	moduleLoader->UnloadModules();
-}
-
-
-// TApp is an exported class from Outpost2.exe.
-// We need to replace some of its methods with a compatible signature
-class TApp
-{
-public:
-	int Init();
-	void ShutDown();
-};
-
-// Declaration for patch to LoadLibrary, where it loads OP2Shell.dll
-// Must use WINAPI macro (__stdcall specifier) to ensure callee cleans the stack
-// By default, for plain functions, the caller cleans the stack, rather than the callee
-HINSTANCE WINAPI LoadShell(LPCSTR lpLibFileName);
-
-DWORD* loadLibraryDataAddr = (DWORD*)0x00486E0A;
-DWORD loadLibraryNewAddr = (DWORD)LoadShell;
-
-
-bool InstallTAppEventHooks()
-{
-	// Replace call to gTApp.Init with custom routine
-	if (!Op2RelinkCall(0x004A8877, GetMethodVoidPointer(&TApp::Init))) {
-		return false;
-	}
-
-	// Replace call to gTApp.ShutDown with custom routine
-	Op2RelinkCall(0x004A88A5, GetMethodVoidPointer(&TApp::ShutDown));
-
-	// Replace call to LoadLibrary with custom routine (address is indirect)
-	Op2MemSetDword(loadLibraryDataAddr, &loadLibraryNewAddr);
-
-	return true;
-}
-
-
-int TApp::Init()
-{
-	// Trigger event
-	OnInit();
-
-	// Call original function
-	return (this->*GetMethodPointer<decltype(&TApp::Init)>(0x00485B20))();
-}
-
-void TApp::ShutDown()
-{
-	// Call original function
-	(this->*GetMethodPointer<decltype(&TApp::ShutDown)>(0x004866E0))();
-
-	// Trigger event
-	OnShutdown();
-}
-
-HINSTANCE WINAPI LoadShell(LPCSTR lpLibFileName)
-{
-	// First try to load it
-	HINSTANCE hInstance = LoadLibraryA(lpLibFileName);
-
-	if (hInstance)
-	{
-		OnLoadShell();
-	}
-
-	return hInstance;
 }
